Needs fewer http connections, manages cookies automatically: https://requests.kennethreitz.org//en/master/user/advanced/#session-objects

*Very* lightly tested, basically just that `tests/dump.py` still appears to work. And it's a bit faster, too.

Before:
```
real	0m4,923s
user	0m1,272s
sys	0m0,115s
```
After:
```
real	0m4,083s
user	0m1,085s
sys	0m0,087s
```
(Both of the above "real" times include a ~1.2 second stall in `syslog.querylog`, so in most usage scenarios the relative difference is even larger than in the above.)