@scop 2019-10-06
Well, most usage scenarios don't hit every possible method of the API though :)