@Salamek 2019-10-06
@scop :smile: i have same chages in my local branch :smile: I just did not want to push them since i have no way to test it... I need to get unit test working...
THX, merging!